Re-jigged the resources
Moved the resources folder out of the ./src folder (leaving that just for .ccp and .h files). Updated the resource .qrc file, and moved it out of the ./src folder to the project root. Deleted the pre-complied icon and edit files from source, no need for them as we re-complie the resources every time.