Moved SPV writer functions in new SPVWriter class
Moved the SPV file writting functions into their own class - out of the SPVreader class. This is to 1) make it easy to find the SPV writing code; and 2) remove all writting code from a class name "reader"! A new class call StaticFunctions has been created to hold helper functions that are required in multiple places but are self contained.