Re-worked scale grid to allow scales to be shown at m, cm, mm, and um
The scale grid will now show scale values in m, cm, mm, or um depending on the FOV and mmToUnit values. An option has been added in the Scale menu to turn on Minor Grid lines values for very very small objects.

One issue that to be resolved is the excessive density of the minor grid lines at the zoomed out cm scale. May be need a step imposed at over 100mm to force scaling to 10's cm. Currently it is showing minor lines every 1 cm!