Tweeked the Output > Settings tab so it fits better on all screen sizes
Changed the frid layout for the settings to only have two columns, not four, as it was cutting off text with the default tab width on my home desktop. This cuttign off issues was almost certainly caused when I increased the decimal and maxium numbers allowed for the Pixles/mm and Slices/mm input boxes.