Added global debug to file functions to SPIERSview
The standard QT debug, warning, info, and fatal macro are caught by the new function in main.cpp. These are either feed only to the console, or to the console and a file named SPIERSView_Debug.log in the users HOME folder. There is a bool flag on line 36 of the main.ccp file that swtiches this behaviour. Also to get this to work in release mode (for debuging release macOS applications) lines 12-14 in the .pro fiule need to be uncommented. Both this and the writting to log file should be turned off in real releases.