Added custom Info.plist to SPIERSview for macOS
macOS file associations are done through the Info.plist files under the application pakages. QT/qmake should be able to overwrite the default generated file with this custom file. Need to make sure that the custom Info.plist (below) is in XML format and not binary as qmake uses sed for string replacements within it.