Fixed non closing of SPIERSview application on macOS when file is not dirty
This fixes the issue where the SPIERSview application stays open in macOS after pressing the 'Red Cross' but were the GUI is removed. This needs testing on Windows to see if it is going to cause problems... if so some IF definition staments can be added to make it macOS only.