Fixed SPIERSedit opening on file double click on macOS
Note that this now opens a new SPIERSedit application thn loads the selected file; previous behaviour was (I think) to open the slected file in the already open window. There is little downside to changing this, however it is now posible to open the same SPIERSedit settings file twice in different SPIERSedit applciation on the same machine via the double click.